+++ src/main/java/com/swansoftwaresolutions/jirareport/core/service/impl/DashboardServiceImpl.java
@@ -73,7 +73,7 @@ public class DashboardServiceImpl implements DashboardService {
             projectReportDto.setActualUatDefectHours(total.getuActualHours());
             projectReportDto.setChart(new Chart(
                     total.getChartLabels().split(","),
-                    Arrays.stream(total.getChartTarget().substring(1, total.getChartTarget().length()-1).split(",")).map(String::trim).mapToInt(Integer::parseInt).toArray(),
+                    Arrays.stream(total.getChartTarget().substring(1, total.getChartTarget().length()-1).split(",")).map(String::trim).mapToDouble(Double::parseDouble).toArray(),
                     Arrays.stream(total.getChartActual().substring(1, total.getChartActual().length()-1).split(",")).map(String::trim).mapToInt(Integer::parseInt).toArray()
             ));
